+++ certbot/certbot/tests/util.py
@@ -16,7 +16,7 @@ from typing import Iterable
 from typing import List
 from typing import Optional
 import unittest
-import warnings
+from unittest import mock
 
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
@@ -35,25 +35,11 @@ from certbot.compat import os
 from certbot.display import util as display_util
 from certbot.plugins import common
 
-try:
-    # When we remove this deprecated import, we should also remove the
-    # "external-mock" test environment and the mock dependency listed in
-    # tools/pinning/pyproject.toml.
-    import mock
-    warnings.warn(
-        "The external mock module is being used for backwards compatibility "
-        "since it is available, however, future versions of Certbot's tests will "
-        "use unittest.mock. Be sure to update your code accordingly.",
-        PendingDeprecationWarning
-    )
-except ImportError:  # pragma: no cover
-    from unittest import mock  # type: ignore
-
 
 class DummyInstaller(common.Installer):
     """Dummy installer plugin for test purpose."""
     def get_all_names(self) -> Iterable[str]:
-        pass
+        return []
 
     def deploy_cert(self, domain: str, cert_path: str, key_path: str, chain_path: str,
                     fullchain_path: str) -> None:
@@ -64,7 +50,7 @@ class DummyInstaller(common.Installer):
         pass
 
     def supported_enhancements(self) -> List[str]:
-        pass
+        return []
 
     def save(self, title: Optional[str] = None, temporary: bool = False) -> None:
         pass
@@ -83,7 +69,7 @@ class DummyInstaller(common.Installer):
         pass
 
     def more_info(self) -> str:
-        pass
+        return ""
 
 
 def vector_path(*names: str) -> str:
@@ -153,7 +139,7 @@ def load_pyopenssl_private_key(*names: str) -> crypto.PKey:
     return crypto.load_privatekey(loader, load_vector(*names))
 
 
-def make_lineage(config_dir: str, testfile: str, ec: bool = False) -> str:
+def make_lineage(config_dir: str, testfile: str, ec: bool = True) -> str:
     """Creates a lineage defined by testfile.
 
     This creates the archive, live, and renewal directories if
@@ -198,56 +184,18 @@ def make_lineage(config_dir: str, testfile: str, ec: bool = False) -> str:
     return conf_path
 
 
-def patch_get_utility(target: str = 'zope.component.getUtility') -> mock.MagicMock:
-    """Deprecated, patch certbot.display.util directly or use patch_display_util instead.
-
-    :param str target: path to patch
-
-    :returns: mock zope.component.getUtility
-    :rtype: mock.MagicMock
-
-    """
-    warnings.warn('Decorator certbot.tests.util.patch_get_utility is deprecated. You should now '
-                  'patch certbot.display.util yourself directly or use '
-                  'certbot.tests.util.patch_display_util as a temporary workaround.')
-    return cast(mock.MagicMock, mock.patch(target, new_callable=_create_display_util_mock))
-
-
-def patch_get_utility_with_stdout(target: str = 'zope.component.getUtility',
-                                  stdout: Optional[IO] = None) -> mock.MagicMock:
-    """Deprecated, patch certbot.display.util directly
-    or use patch_display_util_with_stdout instead.
-
-    :param str target: path to patch
-    :param object stdout: object to write standard output to; it is
-        expected to have a `write` method
-
-    :returns: mock zope.component.getUtility
-    :rtype: mock.MagicMock
-
-    """
-    warnings.warn('Decorator certbot.tests.util.patch_get_utility_with_stdout is deprecated. You '
-                  'should now patch certbot.display.util yourself directly or use '
-                  'use certbot.tests.util.patch_display_util_with_stdout as a temporary '
-                  'workaround.')
-    stdout = stdout if stdout else io.StringIO()
-    freezable_mock = _create_display_util_mock_with_stdout(stdout)
-    return cast(mock.MagicMock, mock.patch(target, new=freezable_mock))
-
-
 def patch_display_util() -> mock.MagicMock:
     """Patch certbot.display.util to use a special mock display utility.
 
     The mock display utility works like a regular mock object, except it also
     also asserts that methods are called with valid arguments.
 
-    The mock created by this patch mocks out Certbot internals so this can be
-    used like the old patch_get_utility function. That is, the mock object will
-    be called by the certbot.display.util functions and the mock returned by
-    that call will be used as the display utility. This was done to simplify
-    the transition from zope.component and mocking certbot.display.util
-    functions directly in test code should be preferred over using this
-    function in the future.
+    The mock created by this patch mocks out Certbot internals. That is, the
+    mock object will be called by the certbot.display.util functions and the
+    mock returned by that call will be used as the display utility. This was
+    done to simplify the transition from zope.component and mocking
+    certbot.display.util functions directly in test code should be preferred
+    over using this function in the future.
 
     See https://github.com/certbot/certbot/issues/8948
 
@@ -267,13 +215,12 @@ def patch_display_util_with_stdout(
     The mock display utility works like a regular mock object, except it also
     asserts that methods are called with valid arguments.
 
-    The mock created by this patch mocks out Certbot internals so this can be
-    used like the old patch_get_utility function. That is, the mock object will
-    be called by the certbot.display.util functions and the mock returned by
-    that call will be used as the display utility. This was done to simplify
-    the transition from zope.component and mocking certbot.display.util
-    functions directly in test code should be preferred over using this
-    function in the future.
+    The mock created by this patch mocks out Certbot internals. That is, the
+    mock object will be called by the certbot.display.util functions and the
+    mock returned by that call will be used as the display utility. This was
+    done to simplify the transition from zope.component and mocking
+    certbot.display.util functions directly in test code should be preferred
+    over using this function in the future.
 
     See https://github.com/certbot/certbot/issues/8948
 
@@ -306,7 +253,7 @@ class FreezableMock:
     value of func is ignored.
 
     """
-    def __init__(self, frozen: bool = False, func: Callable[..., Any] = None,
+    def __init__(self, frozen: bool = False, func: Optional[Callable[..., Any]] = None,
                  return_value: Any = mock.sentinel.DEFAULT) -> None:
         self._frozen_set = set() if frozen else {'freeze', }
         self._func = func