+++ certbot/tests/reverter_test.py
@@ -4,11 +4,7 @@ import logging
 import shutil
 import tempfile
 import unittest
-
-try:
-    import mock
-except ImportError: # pragma: no cover
-    from unittest import mock
+from unittest import mock
 
 from certbot import errors
 from certbot.compat import os